feat!: Update to use Eigen3 inputs and outputs
The Hungarian code has been updated to use Google's version as a base, which
was very similar to the original code but works for rectangular matrices where
the original implementation didn't. The Hungarian code has also been
modified to utilize Eigen3 inputs and outputs. The `HungarianOptimizer`
constructor takes in an Eigen::Matrix, which it utilizes to set up the necessary
internal data used during the `Minimize()` and `Maximize()` process. The
`Maximize()` and `Minimize()` functions have been updated to not take in any
parameters, but instead will output an `std::pair` of Eigen::Vectors. The first
pair of Eigen::Vectors will contain the row indices corresponding to the optimal
cost found for the job assignment solution, while the second pair will contain
the column indices for this solution.

A static convenience function in `GetTotalCost()` has been added to get the
total cost associated with a given cost matrix and its corresponding row/column
indices that correspond to the optimal solution. The `PrintMatrix()` function
is now static and accepts the same parameters to print the matrix with the
values/cells corresponding to the optimal solution prepended with a `*`.

The sample `main.cpp` now utilizes DocTest to run the tests provided in the
new `TestHungarian.cpp` file. These tests include testing the accuracy of the
algorithm with various cost matrices, as well as some performance tests for
matrices of `100x100`, `1000x1000`, and `10000x10000` sizes. These tests can
also be used as examples for functionality usage.

The dependencies and instructions have been updated to reflect these updates.